# BACKEND/docs/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Vizsgaremek Server API Dokumentation
  description: Detailed description and use of Vizsgaremek's paths
servers:
  - url: http://localhost:3000
paths:
...
  /users:
    get:
      description: Return a list of users
      security: 
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content: 
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/User'
    post:
      description: Creates a new user
      security: 
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - message
        - has_error
      properties:
        message:
          type: string
        has_error:
          type: boolean